include tests for force user defined cleanup
